html/semantics/embedded-content/media-elements/track/track-element/track-api-length.html

<!DOCTYPE HTML>
<html>
<head>
<title>Video and Audio text tracks API test</title>
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/embedded-content-0.html#text-track-api">
<link rel="author" title="Sungmin Jang">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Media sample</h1>
<div style="display:none">
	<video id="tracktest" src="brave.webm">
		<track kind=subtitles src=brave.en.vtt srclang=en label="English">
		<track kind=captions src=brave.en.vtt srclang=en label="English for the Hard of Hearing">
		<track kind=subtitles src=brave.fr.vtt srclang=fr label="Français">
		<track kind=subtitles src=brave.de.vtt srclang=de label="Deutsch">
	</video>
</div>
<div id="log"></div>
<script>
var t = document.getElementById('tracktest').textTracks;
function checkTracks() {
	if(t) return true;
}
test(function() {
	assert_true(checkTracks());
	assert_equals(4,t.length);
}, "Check the track elements, count tracks");
</script>
</body>
</html>
